test: integration test base
